use random to generate initial graph, cliquesize us long

+++ TcpClient.java
@@ -36,7 +36,7 @@ public class TcpClient {
     private String destHost;
     private int destPort;
     private int currentSize = 0;
-    private int cliqueSize = Integer.MAX_VALUE;
+    private long cliqueSize = Long.MAX_VALUE;
     private String currentGraph = " ";
     private String backupAddr = " ";
     private int backupPort = -1;
@@ -53,7 +53,7 @@ public class TcpClient {
         return currentSize;
     }
 
-    public int getCliqueSize() {
+    public long getCliqueSize() {
         return cliqueSize;
     }
 
@@ -87,18 +87,18 @@ public class TcpClient {
         backupAddr = read();
         backupPort = Integer.parseInt( read() );
         currentSize = Integer.parseInt( read() );
-        cliqueSize = Integer.parseInt( read() );
+        cliqueSize = Long.parseLong( read() );
         currentGraph = read();
         System.out.println( "Client start to work on problem with size " +
                             Integer.toString( currentSize ) + " and clique size " +
-                            Integer.toString( cliqueSize ) );
+                            Long.toString( cliqueSize ) );
 
     }
 
     // call by algorithm, start the exchange with server
     // if graph from update is invalid, graph will be empty
     // the reuslt graph return by getter will be all -1
-    public void updateFromAlg( int problemSize, int cliqueSize, int[][] graph ) {
+    public void updateFromAlg( int problemSize, long cliqueSize, int[][] graph ) {
         this.currentSize = problemSize;
         this.cliqueSize = cliqueSize;
         this.currentGraph = translateGraphToString( graph );
@@ -163,7 +163,7 @@ public class TcpClient {
         backupAddr = read();
         backupPort = Integer.parseInt( read() );
         write( new String[] { Integer.toString( currentSize ),
-                              Integer.toString( cliqueSize )} );
+                              Long.toString( cliqueSize )} );
         message = read();
         if ( message.equals( requestMessage ) ) {
             // case A_0 and case A_1
@@ -203,7 +203,7 @@ public class TcpClient {
     public void handleProblemSizeChanged() {
         String message = "";
         this.currentSize = Integer.parseInt( read() );
-        this.cliqueSize = Integer.parseInt( read() );
+        this.cliqueSize = Long.parseLong( read() );
         this.currentGraph = read();
         System.out.println( "problem size not matched with server, now = " + this.currentSize );
         message = read();
@@ -224,7 +224,7 @@ public class TcpClient {
         if ( message.equals( tieMessage ) ) {
             System.out.println( "server and client haave same clique" );
         } else {
-            this.cliqueSize = Integer.parseInt( message );
+            this.cliqueSize = Long.parseLong( message );
             this.currentGraph = read();
             System.out.println( "server has better clique " + message );
         }
